refactor(Live Stage): separate game pieces control to its own bloc
Refactor pieces so that they don't have game logic.
Make a live stage bloc structure to handle the pieces,
so that we can handle when a drag piece could go to two target pieces.
All possible targets will warn and may accept the drag,
but only one will take it when the user has no more attempts.

This Live Stage bloc is a 'child bloc' of the game bloc,
which now only handles the logic for changing levels and stages
and with the level/stage information creates the new live stage
bloc to handle the game.

In the way, added safe area consideration and fixed positions
calculation based on that.